Refactor edition detection for Ember RFC #558.
Adds support for the system laid out in Ember RFC #558, in a backwards
compatible way.

Summary of changes:

* `setEdition` and `clearEdition` are "soft" deprecated
* `has` now accepts a `projectRoot` path as a second argument which
  defaults to the processes current working directory
* `has` will check the `projectRoot` for a `package.json`, and if it
  contains `ember.edition` key that will be returned
* `has` continues to fall back to leveraging the preexisting environment
  variables if `package.json` was not found or if the package.json does
  not contain an `ember.edition` property.

Note: this commit creates a "soft deprecation" so that the ecosystem has
a window of time in order to migrate away from `setEdition` without
significant console noise. After a short time period has elapsed with
the new mechanism introduced, we will begin issue console warnings for
usage of `setEdition`.